requirements.txt
PySide6
pyflowreg
pyyaml
tomli; python_version < '3.11'
